pyproject.toml
[project]
name = "service-agent-voice"
version = "0.1.0"
description = "Gateway service for STT, translation, and TTS orchestration."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "grpcio>=1.78.0",
    "protobuf>=5.0",
    "requests>=2.31",
    "uvicorn[standard]>=0.29",
    "websockets>=15.0.1",
]

[project.optional-dependencies]
desktop = [
    "numpy>=1.24",
    "sounddevice>=0.4.6",
]
dev = [
    "grpcio-tools>=1.78.0",
]

[tool.uv]
package = false
